Repairing the skin

+++ 0x02-python-import_modules/calculator_1.py
@@ -46,4 +46,4 @@ def div(a, b):
     Returns:
         The return value. a / b
     """
-    return (a / b)
+    return int(a / b)
